Broke up create_regexp() into smaller parts.
The functionality of create_regexp() is now divided into two smaller
functions, create_regex() and create_regexlist().
The former just creates a regex from a pattern string, while the latter
arranges this output in a regexlist data structure.

This was done for refactoring; I'll later need the bare functionality of
create_regex().